Add proper exception breakpoint to catch exceptions occurred during stream tracing
Change breakpoint requestors behaviour, so they are able to pause execution
Time ticking now optional in collectors (made for compatibility with result interpreters)
Fix tests
