+++ plugins/stream-debugger/support-lib/src/com/intellij/debugger/stream/rt/java/collectors/DoubleCollector.java
@@ -8,14 +8,16 @@ package com.intellij.debugger.stream.rt.java.collectors;
 class DoubleCollector implements java.util.function.DoubleConsumer {
     private final java.util.Map<Integer, Object> storage;
     private final java.util.concurrent.atomic.AtomicInteger time;
+    private final boolean tick;
 
-    DoubleCollector(java.util.Map<Integer, Object> storage, java.util.concurrent.atomic.AtomicInteger time) {
+    DoubleCollector(java.util.Map<Integer, Object> storage, java.util.concurrent.atomic.AtomicInteger time, boolean tick) {
         this.storage = storage;
         this.time = time;
+        this.tick = tick;
     }
 
     @Override
     public void accept(double t) {
-        storage.put(time.incrementAndGet(), t);
+        storage.put(tick ? time.incrementAndGet() : time.get(), t);
     }
 }
